project
full project

+++ src/App/AdminBundle/Controller/PackagesController.php
@@ -9,15 +9,30 @@ use Symfony\Component\HttpFoundation\Request;
 
 class PackagesController extends Controller
 {
+
+    public $adminloggedin = false;
+    public $admin = null;
+
+    public function __construct()
+    {
+        if (isset($_SESSION['admin'])){
+            $this->admin = $_SESSION['admin'];
+            $this->adminloggedin = true;
+        }
+    }
+
     /**
      * @Route("/Packages", name="admin_packages")
      */
     public function PackagesAction()
     {
-        $package = $this->getDoctrine()->getRepository(Package::class)->findAll();
-        return $this->render('AdminBundle:Admin/Packages:packages.html.twig', array(
-            'Packages' => $package
-        ));
+        if ($this->adminloggedin == true){
+            $package = $this->getDoctrine()->getRepository(Package::class)->findAll();
+            return $this->render('AdminBundle:Admin/Packages:packages.html.twig', array(
+                'Packages' => $package
+            ));
+        }
+        return $this->redirectToRoute('workplace_login');
     }
 
     /**
@@ -25,18 +40,21 @@ class PackagesController extends Controller
  */
     public function AddPackagesAction(Request $request)
     {
-        if ($request->getMethod() == "POST"){
-            $em = $this->getDoctrine()->getManager();
-            $pkg = new Package();
-            $pkg->setName($request->get('name'));
-            $pkg->setPrice($request->get('price'));
-            $pkg->setMonth($request->get('month'));
-            $pkg->setInterest($request->get('interest'));
-            $em->persist($pkg);
-            $em->flush();
-            return $this->redirectToRoute('admin_packages');
+        if ($this->adminloggedin == true){
+            if ($request->getMethod() == "POST"){
+                $em = $this->getDoctrine()->getManager();
+                $pkg = new Package();
+                $pkg->setName($request->get('name'));
+                $pkg->setPrice($request->get('price'));
+                $pkg->setMonth($request->get('month'));
+                $pkg->setInterest($request->get('interest'));
+                $em->persist($pkg);
+                $em->flush();
+                return $this->redirectToRoute('admin_packages');
+            }
+            return $this->render('AdminBundle:Admin/Packages:addpackages.html.twig');
         }
-        return $this->render('AdminBundle:Admin/Packages:addpackages.html.twig');
+        return $this->redirectToRoute('workplace_login');
     }
 
     /**
@@ -44,20 +62,23 @@ class PackagesController extends Controller
      */
     public function UpdatePackagesAction(Request $request, $id)
     {
-        if ($request->getMethod() == "POST"){
-            $em = $this->getDoctrine()->getManager();
-            $pkg = $em->getRepository(Package::class)->find($id);
-            $pkg->setName($request->get('name'));
-            $pkg->setPrice($request->get('price'));
-            $pkg->setMonth($request->get('month'));
-            $pkg->setInterest($request->get('interest'));
-            $em->flush();
-            return $this->redirectToRoute('admin_packages');
+        if ($this->adminloggedin == true){
+            if ($request->getMethod() == "POST"){
+                $em = $this->getDoctrine()->getManager();
+                $pkg = $em->getRepository(Package::class)->find($id);
+                $pkg->setName($request->get('name'));
+                $pkg->setPrice($request->get('price'));
+                $pkg->setMonth($request->get('month'));
+                $pkg->setInterest($request->get('interest'));
+                $em->flush();
+                return $this->redirectToRoute('admin_packages');
+            }
+            $pkg1 = $this->getDoctrine()->getRepository(Package::class)->find($id);
+            return $this->render('AdminBundle:Admin/Packages:updatepackages.html.twig', array(
+                'Package' => $pkg1
+            ));
         }
-        $pkg1 = $this->getDoctrine()->getRepository(Package::class)->find($id);
-        return $this->render('AdminBundle:Admin/Packages:updatepackages.html.twig', array(
-            'Package' => $pkg1
-        ));
+        return $this->redirectToRoute('workplace_login');
     }
 
     /**
@@ -65,11 +86,14 @@ class PackagesController extends Controller
      */
     public function DeletePackagesAction($id)
     {
+        if ($this->adminloggedin == true){
             $em = $this->getDoctrine()->getManager();
             $pkg = $em->getRepository(Package::class)->find($id);
             $em->remove($pkg);
             $em->flush();
             return $this->redirectToRoute('admin_packages');
+        }
+        return $this->redirectToRoute('workplace_login');
     }
